backend/requirements.txt
fastapi
sqlmodel
python-dotenv
pygithub
httpx
psycopg2-binary
redis
uvicorn
python-jose[cryptography]
google-genai
